Unify FileSystem file-descriptor methods behind common bases (JNA/FFM dedup)
First subsystem in the cross-OS consistency audit. Hoist the file-descriptor
query methods to common base classes so the JNA/FFM subclasses shrink to a
single native hook, matching the pattern FreeBsdFileSystem already used:

- macOS: MacFileSystem base gains the three FD methods + a querySysctl(String)
  hook; MacFileSystemJNA/FFM drop to the one-line sysctl call.
- OpenBSD: same treatment on OpenBsdFileSystem.
- Windows: new WindowsFileSystem base holds MAX_WINDOWS_HANDLES + the FD methods
  behind a queryHandles() hook; JNA/FFM extend it.

No user-facing behavior change. The FFM Windows handle limit had hardcoded the
64-bit value while JNA computed it conditionally, but FFM runs only on 64-bit
Windows (JDK 25+) where the values coincide; the shared base uses the
conditional form to keep both backends consistent.

The native getFileStores volume enumeration stays per-backend (Kernel32/WMI vs
getfsstat are genuinely different native code); its JNA/FFM merge-loop drift is
deferred to a follow-up (coupled to the static WindowsOSFileStore
updateAttributes calls).
